% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{get_meanPurityCorr}
\alias{get_meanPurityCorr}
\title{Correlation between gene expression and purity by TAD}
\usage{
get_meanPurityCorr(exprTable, purityTable, g2tTable, all_samples,
  purityCol, sampleCol = "Sample.ID", transfExpr = "log10",
  logOffset = 0.01, corrMeth = "pearson", nCpu = 2)
}
\arguments{
\item{exprTable}{Dataframe with gene expression values for which to compute correlation (row names should be gene IDs; all gene IDs from g2tTable should be available).}

\item{purityTable}{Dataframe with sample purity values (required columns: purityCol and sampleCol).}

\item{g2tTable}{Dataframe with gene-to-TAD assignment (required columns: "entrezID" for gene IDs, "region" for TAD IDs).}

\item{all_samples}{Vector of samples. Should match exprTable columns and sampleCol column of purityTable.}

\item{purityCol}{Name of the column of purityTable that holds purity values.}

\item{sampleCol}{Name of the column of purityTable that holds sample IDs.}

\item{transfExpr}{Transformation to apply to gene expression before computing gene expression (should be a function name, such as log10).}

\item{logOffset}{Numeric value added to gene expression before applying transfExpr (might be useful for log-transformation).}

\item{corrMeth}{Correlation method (one of "pearson", "kendall"or "spearman"; default:"pearson").}

\item{nCpu}{Number available CPU.}
}
\value{
A 4-column dataframe (nSampWithPurity/region/entrezID/purityCorr colums). The returned dataframe is not averaged at TAD level.
}
\description{
Function the correlation between gene expression and purity for all TADs.
}
